+++ tests/api_test.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8
 """Unit tests for didyoumean APIs."""
-from didyoumean_api import didyoumean_decorator, didyoumean_contextmanager,\
+from didyoumean.api import didyoumean_decorator, didyoumean_contextmanager,\
     didyoumean_postmortem, didyoumean_enablehook, didyoumean_disablehook
-from didyoumean_common_tests import TestWithStringFunction,\
+from tests.common_test import TestWithStringFunction,\
     get_exception, no_exception, NoFileIoError, unittest_module
 import contextlib
 import sys
@@ -266,7 +266,7 @@ class IPythonHookTest(unittest_module.TestCase, HookTest):
         """Run code with didyoumean after enabling didyoumean hook."""
         prev_handler = None
         shell = DummyShell()
-        module = sys.modules['didyoumean_api']
+        module = sys.modules['didyoumean.api']
         shell.set(module)
         self.assertEqual(shell.handler, prev_handler)
         didyoumean_enablehook()
